requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0
reportlab==4.0.4
validators==0.22.0
tldextract==5.1.1
beautifulsoup4==4.12.2
pillow==10.0.1
qrcode==7.4.2
lxml==4.9.3
